General GUI and Text Input
Configured a significant portion of the Swing components. Four different
sortPanels (subclasses of JPanel) added to the Frame for selection,
insertion, merge, and quick sort. Buttons and instructions added to take
values from adjacent JTextPane.

Method added which parses a String (in general from JTextPane) and
parses all positive integers using regex to prevent exceptions from
being thrown. The valid numbers in the String are added to each of the
four sortPanels and then everything is repainted.

Ideas/To do:
-code the algorithms themselves
-configure a delay (and perhaps a pause button) to show steps
-add JLabels to each sortPanel describing the instruction/step
-add a colour array to sortPanel to indicate what color to make each bar
based on state (sorted, unsorted, greater than, less than, not relevant
at the moment, etc.)
-have the system run the sorts once without delay to compare time
elapsed
-consider removing start button and having each sortPanel start/stop
when clicked on. (maybe add a MouseListener to each sortPanel either in
MainWindow or in the class definition if possible, then a mouse click
will invoke the listener and sleep until another click is done)